small changes and lots of files re-organization
